Move this repo to flat structure (classic)
