+++ addon/components/frost-tab.js
@@ -1,5 +1,5 @@
 import Ember from 'ember'
-import layout from './template'
+import layout from '../templates/components/frost-tab'
 import PropTypesMixin, { PropTypes } from 'ember-prop-types'
 
 const {
