+++ addon/templates/components/frost-tab.hbs
