+++ addon/templates/components/frost-tabs.hbs
