+++ app/components/frost-tab.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-frost-tabs/pods/components/frost-tab/component'
+export { default } from 'ember-frost-tabs/components/frost-tab'
