+++ app/components/frost-tabs.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-frost-tabs/pods/components/frost-tabs/component'
+export { default } from 'ember-frost-tabs/components/frost-tabs'
