+++ config/environment.js
@@ -1,7 +1,5 @@
 'use strict'
 
 module.exports = function (/* environment, appConfig */) {
-  return {
-    podModulePrefix: 'ember-frost-tabs/pods'
-  }
+  return { }
 }
