+++ package.json
@@ -10,6 +10,7 @@
   "scripts": {
     "start": "ember server",
     "build": "ember build",
+    "deploy": "deploy.sh",
     "test": "npm run lint && COVERAGE=true ember test",
     "lint": "npm run eslint && npm run sass-lint && npm run md-lint",
     "md-lint": "remark *.md",
@@ -53,7 +54,7 @@
     "ember-elsewhere": "0.4.1",
     "ember-export-application-global": "^1.0.5",
     "ember-frost-core": "^0.29.0",
-    "ember-hook": "1.3.5",
+    "ember-hook": "^1.3.5",
     "ember-load-initializers": "^0.5.1",
     "ember-lodash-shim": "1.0.1",
     "ember-one-way-controls": "1.1.1",
